Create README.md
Sun moon coin sun head moon tails/ value to be voted on
EnvoEcovalue digital device. A bluetoother device u buy were u down load a app if you loose your phone or device dont worry it bluetooth connected all your information with the enter of a code is erased and saved on the device that the code was entered to last...
   DiversityDiversifydigitalphysicalcash coins. And physical tokens with Indians and early American settlers history changing people on the tokens the are cashed at store from face value also on line by code that's on them the way they u can tell if there spent is buy the code that can be scanned there reusable refillable. A card no a coin made from scrapp aluminium also able to bring the scrap metal price up. Made from all scrap metals it just depends face value u purchase.. the way to fight counterfeit is up to the dor. And there mark. I want Native Americans On them. White and African world leaders from history and currently immigrate and foriegn appreciation  physicalcashcoins. Also for charity and donations depending on if it be For God or to a religious place of worship I want in God we trust and believe forever and always with The cross and the world in his hands. 
 You the franchise or organization pay to have your own reloadable coins made. U choose design
The Goverment for taxes and whatever makes them all but has there own face and type
The  citizens communities vote or on there designs.
There is one unicurrency coin and its I think if we can get other countries to agree is acceptable as payment everywhere.